Get Mobile Android Java example working again (#182)
- A few fixes:
  - Checkpoint was changed from a directory to a single file
  - Type of labels input in the training model was changed to int64
- Remove checked in training artifact assets, add helper script to copy artifacts to app assets directory
- Added a basic test that creates a training session

+++ on_device_training/mobile/android/java/app/ORTPersonalize/app/src/main/java/com/example/ortpersonalize/ORTTrainer.kt
@@ -8,6 +8,7 @@ import androidx.annotation.InspectableProperty.ValueType
 import java.io.File
 import java.nio.FloatBuffer
 import java.nio.IntBuffer
+import java.nio.LongBuffer
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.util.Collections
@@ -24,7 +25,7 @@ class ORTTrainer{
         ortTrainingSession = ortEnv?.createTrainingSession(checkpointPath, trainModelPath, evalModelPath, optimizerModelPath)
     }
 
-    public fun performTraining(data: FloatBuffer, labels: IntBuffer, batchSize: Long) {
+    public fun performTraining(data: FloatBuffer, labels: LongBuffer, batchSize: Long) {
         ortSession = null
 
         var loss = -1.0f
